refactor(chart): revert back import path update
